+++ NEWS.md
@@ -1,6 +1,7 @@
 # musicMCT (development version)
 
 * `vl_generators()` now gives correct results for sets which fail `optc_test()`
+* `minimize_vl()` now returns better results when `method="hamming"` by allowing for voice crossings (#4).
 
 # musicMCT 0.2.0